backend/requirements.txt
pdf2image
pytesseract
opencv_python
numpy
pandas
pillow
fastapi[all]
pytest
